ai-service/requirements.txt
openai>=1.0.0
spacy>=3.7.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
